+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com/appwise-labs/AppwiseCore)
 
-* Fastlane: import localization files sorted by name
+### Improvements
+
+* Core: `Router` now supports an `anyParams` property which you can use with `JSONEncoding` to send (for example) an `Array` instead of a `Dictionary`.
+
+### Bug Fixes
+
+* Fastlane: import localization files sorted by name.
 
 ## [1.3.6](https://github.com/appwise-labs/AppwiseCore/releases/tag/1.3.6)